+++ examples/cryptocurrency-advanced/frontend/src/pages/Block.vue
@@ -74,6 +74,7 @@
       return {
         block: {},
         transactions: [],
+        isSpinnerVisible: false
       }
     },
     computed: {