+++ exonum/src/blockchain/service.rs
@@ -18,10 +18,10 @@
 use iron::Handler;
 use serde_json::Value;
 
-use std::collections::{HashMap, HashSet};
-use std::fmt;
-use std::net::SocketAddr;
-use std::sync::{Arc, RwLock};
+use std::{collections::{HashMap, HashSet},
+          fmt,
+          net::SocketAddr,
+          sync::{Arc, RwLock}};
 
 use super::transaction::Transaction;
 use blockchain::{Blockchain, ConsensusConfig, Schema, StoredConfiguration, ValidatorKeys};
@@ -320,9 +320,10 @@ pub struct ApiNodeState {
     incoming_connections: HashSet<SocketAddr>,
     outgoing_connections: HashSet<SocketAddr>,
     reconnects_timeout: HashMap<SocketAddr, Milliseconds>,
-    //TODO: update on event?
+    // TODO: Update on event? (ECR-1632)
     peers_info: HashMap<SocketAddr, PublicKey>,
     is_enabled: bool,
+    majority_count: usize,
 }
 
 impl ApiNodeState {
@@ -395,13 +396,23 @@ impl SharedNodeState {
     }
     /// Updates internal state, from `State` of a blockchain node.
     pub fn update_node_state(&self, state: &State) {
+        let mut lock = self.state.write().expect("Expected write lock.");
+
+        lock.peers_info.clear();
+        lock.majority_count = state.majority_count();
+
         for (p, c) in state.peers().iter() {
-            self.state
-                .write()
-                .expect("Expected write lock.")
-                .peers_info
-                .insert(c.addr(), *p);
+            lock.peers_info.insert(c.addr(), *p);
         }
+    }
+
+    /// Returns the majority count from the current "State"
+    /// of a blockchain node.
+    pub fn majority_count(&self) -> usize {
+        self.state
+            .read()
+            .expect("Expected read lock.")
+            .majority_count
     }
 
     /// Returns a boolean value which indicates whether the node is enabled