+++ exonum/src/encoding/serialize/json.rs
@@ -16,25 +16,19 @@ use bit_vec::BitVec;
 use chrono::{DateTime, Duration, TimeZone, Utc};
 use hex::FromHex;
 use rust_decimal::Decimal;
-/// trait `ExonumSerializeJson` implemented for all field that allows serializing in
-/// json format.
-///
-// TODO refer to difference between json serialization and exonum_json (ECR-156).
-// TODO implement Field for float (ECR-153).
-// TODO remove WriteBufferWrapper hack (after refactor storage),
-// should be moved into storage (ECR-156).
+/// trait `ExonumSerializeJson` implemented for all field that allows serializing in json format.
 use serde_json::{self, value::Value};
 use uuid::Uuid;
 
-use std::error::Error;
-use std::net::SocketAddr;
+use std::{error::Error, net::SocketAddr};
 
 use super::WriteBufferWrapper;
 use crypto::{Hash, PublicKey, Signature};
 use encoding::{Field, Offset};
 use helpers::{Height, Round, ValidatorId};
 use messages::RawMessage;
-// TODO: should we implement serialize for: `SecretKey`, `Seed` (ECR-156)?
+
+// TODO: Should we implement serialize for: `SecretKey`, `Seed`. (ECR-156)
 
 macro_rules! impl_default_deserialize_owned {
     (@impl $name:ty) => {
@@ -366,8 +360,8 @@ where
     }
 }
 
-// TODO remove `ExonumJsonDeserialize` needs
-// after it remove impl `ExonumJsonDeserialize` for all types expect struct (ECR-156)
+// TODO: Remove `ExonumJsonDeserialize` needs
+// after it remove impl `ExonumJsonDeserialize` for all types expect struct. (ECR-156)
 impl<T> ExonumJson for Vec<T>
 where
     T: ExonumJsonDeserialize + ExonumJson,
@@ -434,7 +428,7 @@ impl ExonumJson for BitVec {
     }
 }
 
-// TODO: Make a macro for tuple struct type definitions (ECR-154)?
+// TODO: Make a macro for tuple struct type definitions? (ECR-154)
 impl ExonumJson for Height {
     fn deserialize_field<B: WriteBufferWrapper>(
         value: &Value,