+++ exonum/src/messages/mod.rs
@@ -14,9 +14,9 @@
 
 //! Consensus and other messages and related utilities.
 
-pub use self::protocol::*;
-pub use self::raw::{Message, MessageBuffer, MessageWriter, RawMessage, ServiceMessage,
-                    HEADER_LENGTH, PROTOCOL_MAJOR_VERSION};
+pub use self::{protocol::*,
+               raw::{Message, MessageBuffer, MessageWriter, RawMessage, ServiceMessage,
+                     HEADER_LENGTH, PROTOCOL_MAJOR_VERSION}};
 
 use bit_vec::BitVec;
 
@@ -34,8 +34,8 @@ mod raw;
 #[cfg(test)]
 mod tests;
 
-// TODO: implement common methods for enum types (hash, raw, from_raw, verify)
-// TODO: use macro for implementing enums (ECR-166)
+// TODO: Implement common methods for enum types (hash, raw, from_raw, verify). (ECR-166)
+// TODO: Use macro for implementing enums. (ECR-166)
 
 /// Raw transaction type.
 pub type RawTransaction = RawMessage;