+++ exonum/src/messages/tests.rs
@@ -14,8 +14,7 @@
 
 use crypto::{gen_keypair, PublicKey, SecretKey, Signature};
 use encoding::serialize::FromHex;
-use messages::raw::MessageBuffer;
-use messages::{Message, RawMessage, RawTransaction};
+use messages::{raw::MessageBuffer, Message, RawMessage, RawTransaction};
 
 messages! {
     const SERVICE_ID = 0;