+++ exonum/src/node/requests.rs
@@ -18,9 +18,8 @@ use crypto::{PUBLIC_KEY_LENGTH, SIGNATURE_LENGTH};
 use messages::{BlockRequest, BlockResponse, Message, PrevotesRequest, ProposeRequest,
                RequestMessage, TransactionsRequest, TransactionsResponse, HEADER_LENGTH};
 
-// TODO: height should be updated after any message, not only after status (if signature is correct)
-// TODO: Request propose makes sense only if we know that node is on our height.
-// (ECR-171)
+// TODO: Height should be updated after any message, not only after status (if signature is correct). (ECR-171)
+// TODO: Request propose makes sense only if we know that node is on our height. (ECR-171)
 
 impl NodeHandler {
     /// Validates request, then redirects it to the corresponding `handle_...` function.
@@ -163,10 +162,7 @@ impl NodeHandler {
             msg.from(),
             block,
             precommits.iter().collect(),
-            transactions
-                .iter()
-                .map(|tx_hash| schema.transactions().get(&tx_hash).unwrap())
-                .collect(),
+            &transactions.iter().collect::<Vec<_>>(),
             self.state.consensus_secret_key(),
         );
         self.send_to_peer(*msg.from(), block_msg.raw());