+++ exonum/src/node/state.rs
@@ -18,18 +18,19 @@ use bit_vec::BitVec;
 use failure;
 use serde_json::Value;
 
-use std::collections::{BTreeMap, HashMap, HashSet, hash_map::Entry};
-use std::net::SocketAddr;
-use std::time::{Duration, SystemTime};
+use std::{collections::{hash_map::Entry, BTreeMap, HashMap, HashSet},
+          net::SocketAddr,
+          time::{Duration, SystemTime}};
 
 use blockchain::{ConsensusConfig, StoredConfiguration, ValidatorKeys};
 use crypto::{CryptoHash, Hash, PublicKey, SecretKey};
 use helpers::{Height, Milliseconds, Round, ValidatorId};
-use messages::{Connect, ConsensusMessage, Message, Precommit, Prevote, Propose, RawMessage};
+use messages::{BlockResponse, Connect, ConsensusMessage, Message, Precommit, Prevote, Propose,
+               RawMessage};
 use node::whitelist::Whitelist;
 use storage::{KeySetIndex, MapIndex, Patch, Snapshot};
 
-// TODO: move request timeouts into node configuration (ECR-171)
+// TODO: Move request timeouts into node configuration. (ECR-171)
 
 /// Timeout value for the `ProposeRequest` message.
 pub const PROPOSE_REQUEST_TIMEOUT: Milliseconds = 100;
@@ -84,8 +85,7 @@ pub struct State {
 
     validators_rounds: BTreeMap<ValidatorId, Round>,
 
-    // Current value of the propose timeout.
-    propose_timeout: Milliseconds,
+    incomplete_block: Option<IncompleteBlock>,
 }
 
 /// State of a validator-node.
@@ -102,8 +102,10 @@ pub struct ValidatorState {
 pub enum RequestData {
     /// Represents `ProposeRequest` message.
     Propose(Hash),
-    /// Represents `TransactionsRequest` message.
-    Transactions(Hash),
+    /// Represents `TransactionsRequest` message for `Propose`.
+    ProposeTransactions(Hash),
+    /// Represents `TransactionsRequest` message for `BlockResponse`.
+    BlockTransactions,
     /// Represents `PrevotesRequest` message.
     Prevotes(Round, Hash),
     /// Represents `BlockRequest` message.
@@ -137,6 +139,13 @@ pub struct BlockState {
     patch: Patch,
     txs: Vec<Hash>,
     proposer_id: ValidatorId,
+}
+
+/// Incomplete block.
+#[derive(Clone, Debug)]
+pub struct IncompleteBlock {
+    msg: BlockResponse,
+    unknown_txs: HashSet<Hash>,
 }
 
 /// `VoteMessage` trait represents voting messages such as `Precommit` and `Prevote`.
@@ -242,7 +251,9 @@ impl RequestData {
         #![cfg_attr(feature = "cargo-clippy", allow(match_same_arms))]
         let ms = match *self {
             RequestData::Propose(..) => PROPOSE_REQUEST_TIMEOUT,
-            RequestData::Transactions(..) => TRANSACTIONS_REQUEST_TIMEOUT,
+            RequestData::ProposeTransactions(..) | RequestData::BlockTransactions => {
+                TRANSACTIONS_REQUEST_TIMEOUT
+            }
             RequestData::Prevotes(..) => PREVOTES_REQUEST_TIMEOUT,
             RequestData::Block(..) => BLOCK_REQUEST_TIMEOUT,
         };
@@ -350,6 +361,23 @@ impl BlockState {
     }
 }
 
+impl IncompleteBlock {
+    /// Returns `BlockResponse` message.
+    pub fn message(&self) -> &BlockResponse {
+        &self.msg
+    }
+
+    /// Returns unknown transactions of the block.
+    pub fn unknown_txs(&self) -> &HashSet<Hash> {
+        &self.unknown_txs
+    }
+
+    /// Returns `true` if there are unknown transactions in the block.
+    pub fn has_unknown_txs(&self) -> bool {
+        !self.unknown_txs.is_empty()
+    }
+}
+
 impl State {
     /// Creates state with the given parameters.
     #[cfg_attr(feature = "cargo-clippy", allow(too_many_arguments))]
@@ -402,8 +430,9 @@ impl State {
 
             requests: HashMap::new(),
 
-            propose_timeout: stored.consensus.max_propose_timeout,
             config: stored,
+
+            incomplete_block: None,
         }
     }
 
@@ -495,16 +524,6 @@ impl State {
         trace!("Validator={:#?}", self.validator_state());
 
         self.config = config;
-    }
-
-    /// Sets a new propose timeout value.
-    pub fn set_propose_timeout(&mut self, timeout: Milliseconds) {
-        self.propose_timeout = timeout;
-    }
-
-    /// Returns adjusted (see `TimeoutAdjuster` for the details) value of the propose timeout.
-    pub fn propose_timeout(&self) -> Milliseconds {
-        self.propose_timeout
     }
 
     /// Adds the public key, address, and `Connect` message of a validator.
@@ -702,8 +721,12 @@ impl State {
         self.round.increment();
     }
 
+    /// Return incomplete block.
+    pub fn incomplete_block(&self) -> Option<&IncompleteBlock> {
+        self.incomplete_block.as_ref()
+    }
+
     /// Increments the node height by one and resets previous height data.
-    // FIXME use block_hash
     pub fn new_height(&mut self, block_hash: &Hash, height_start_time: SystemTime) {
         self.height.increment();
         self.height_start_time = height_start_time;
@@ -711,7 +734,7 @@ impl State {
         self.locked_round = Round::zero();
         self.locked_propose = None;
         self.last_hash = *block_hash;
-        // TODO: destruct/construct structure HeightState instead of call clear (ECR-171)
+        // TODO: Destruct/construct structure HeightState instead of call clear. (ECR-171)
         self.blocks.clear();
         self.proposes.clear();
         self.prevotes.clear();
@@ -720,7 +743,8 @@ impl State {
         if let Some(ref mut validator_state) = self.validator_state {
             validator_state.clear();
         }
-        self.requests.clear(); // FIXME: clear all timeouts (ECR-171)
+        self.requests.clear(); // FIXME: Clear all timeouts. (ECR-171)
+        self.incomplete_block = None;
     }
 
     /// Returns a list of queued consensus messages.
@@ -752,6 +776,23 @@ impl State {
         }
 
         full_proposes
+    }
+
+    /// Checks if there is an incomplete block that waits for this transaction.
+    /// Returns a block that don't contain unknown transactions.
+    ///
+    /// Transaction is ignored if the following criteria are fulfilled:
+    ///
+    /// - transaction isn't contained in the unknown transactions list of block
+    /// - transaction isn't a part of block
+    pub fn remove_unknown_transaction(&mut self, tx_hash: Hash) -> Option<IncompleteBlock> {
+        if let Some(ref mut incomplete_block) = self.incomplete_block {
+            incomplete_block.unknown_txs.remove(&tx_hash);
+            if incomplete_block.unknown_txs.is_empty() {
+                return Some(incomplete_block.clone());
+            }
+        }
+        None
     }
 
     /// Returns pre-votes for the specified round and propose hash.
@@ -794,8 +835,9 @@ impl State {
                 propose: msg,
                 unknown_txs: HashSet::new(),
                 block_hash: None,
-                // TODO:: for the moment it's true because this code gets called immediately after
-                // saving a propose to the cache. Think about making this approach less error-prone
+                // TODO: For the moment it's true because this code gets called immediately after
+                // saving a propose to the cache. Think about making this approach less error-prone.
+                // (ECR-1635)
                 is_saved: true,
             },
         );
@@ -804,11 +846,11 @@ impl State {
     }
 
     /// Adds propose from other node. Returns `ProposeState` if it is a new propose.
-    pub fn add_propose(
+    pub fn add_propose<S: AsRef<Snapshot>>(
         &mut self,
         msg: &Propose,
-        transactions: &MapIndex<&&Snapshot, Hash, RawMessage>,
-        transaction_pool: &KeySetIndex<&&Snapshot, Hash>,
+        transactions: &MapIndex<S, Hash, RawMessage>,
+        transaction_pool: &KeySetIndex<S, Hash>,
     ) -> Result<&ProposeState, failure::Error> {
         let propose_hash = msg.hash();
         match self.proposes.entry(propose_hash) {
@@ -863,6 +905,43 @@ impl State {
                 proposer_id,
             })),
         }
+    }
+
+    /// Finds unknown transactions in the block and persists transactions along
+    /// with other info as a pending block.
+    ///
+    ///  # Panics
+    ///
+    /// - Already there is an incomplete block.
+    /// - Received block has already committed transaction.
+    pub fn create_incomplete_block<S: AsRef<Snapshot>>(
+        &mut self,
+        msg: &BlockResponse,
+        txs: &MapIndex<S, Hash, RawMessage>,
+        txs_pool: &KeySetIndex<S, Hash>,
+    ) -> &IncompleteBlock {
+        assert!(self.incomplete_block().is_none());
+
+        let mut unknown_txs = HashSet::new();
+        for hash in msg.transactions() {
+            if txs.get(hash).is_some() {
+                if !txs_pool.contains(hash) {
+                    panic!(
+                        "Received block with already \
+                         committed transaction"
+                    )
+                }
+            } else {
+                unknown_txs.insert(*hash);
+            }
+        }
+
+        self.incomplete_block = Some(IncompleteBlock {
+            msg: msg.clone(),
+            unknown_txs,
+        });
+
+        self.incomplete_block().unwrap()
     }
 
     /// Adds pre-vote. Returns `true` there are +2/3 pre-votes.
@@ -991,7 +1070,7 @@ impl State {
             match self.validator_state {
                 Some(ref validator_state) => {
                     if let Some(msg) = validator_state.our_prevotes.get(&round) {
-                        // TODO: inefficient (ECR-171)
+                        // TODO: Inefficient. (ECR-171)
                         if Some(*msg.propose_hash()) != self.locked_propose {
                             return true;
                         }