+++ exonum/src/sandbox/sandbox.rs
@@ -15,25 +15,24 @@
 // Workaround: Clippy does not correctly handle borrowing checking rules for returned types.
 #![cfg_attr(feature = "cargo-clippy", allow(let_and_return))]
 
-use futures::{self, Async, Future, Sink, Stream, sync::mpsc};
+use futures::{self, sync::mpsc, Async, Future, Sink, Stream};
 
-use std::cell::{Ref, RefCell, RefMut};
-use std::collections::{BTreeSet, BinaryHeap, HashMap, HashSet, VecDeque};
-use std::iter::FromIterator;
-use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-use std::ops::{AddAssign, Deref};
-use std::sync::{Arc, Mutex};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::{cell::{Ref, RefCell, RefMut},
+          collections::{BTreeSet, BinaryHeap, HashMap, HashSet, VecDeque},
+          iter::FromIterator,
+          net::{IpAddr, Ipv4Addr, SocketAddr},
+          ops::{AddAssign, Deref},
+          sync::{Arc, Mutex},
+          time::{Duration, SystemTime, UNIX_EPOCH}};
 
-use super::config_updater::ConfigUpdateService;
-use super::sandbox_tests_helper::VALIDATOR_0;
-use super::timestamping::TimestampingService;
+use super::{config_updater::ConfigUpdateService,
+            sandbox_tests_helper::{VALIDATOR_0, PROPOSE_TIMEOUT},
+            timestamping::TimestampingService};
 use blockchain::{Block, BlockProof, Blockchain, ConsensusConfig, GenesisConfig, Schema, Service,
                  SharedNodeState, StoredConfiguration, Transaction, ValidatorKeys};
 use crypto::{gen_keypair, gen_keypair_from_seed, Hash, PublicKey, SecretKey, Seed};
-use events::network::NetworkConfiguration;
-use events::{Event, EventHandler, InternalEvent, InternalRequest, NetworkEvent, NetworkRequest,
-             TimeoutRequest};
+use events::{network::NetworkConfiguration, Event, EventHandler, InternalEvent, InternalRequest,
+             NetworkEvent, NetworkRequest, TimeoutRequest};
 use helpers::{user_agent, Height, Milliseconds, Round, ValidatorId};
 use messages::{Any, Connect, Message, RawMessage, RawTransaction, Status};
 use node::{ApiSender, Configuration, ExternalMessage, ListenerConfig, NodeHandler, NodeSender,
@@ -232,7 +231,7 @@ impl Sandbox {
 
     pub fn recv<T: Message>(&self, msg: &T) {
         self.check_unexpected_message();
-        // TODO Think about addresses.
+        // TODO Think about addresses. (ECR-1627)
         let dummy_addr = SocketAddr::from(([127, 0, 0, 1], 12_039));
         let event = NetworkEvent::MessageReceived(dummy_addr, msg.raw().clone());
         self.inner.borrow_mut().handle_event(event);
@@ -270,7 +269,7 @@ impl Sandbox {
         self.try_broadcast_to_addrs(msg, self.addresses.iter().skip(1))
     }
 
-    // TODO: add self-test for broadcasting?
+    // TODO: Add self-test for broadcasting? (ECR-1627)
     pub fn broadcast_to_addrs<'a, T: Message, I>(&self, msg: &T, addresses: I)
     where
         I: IntoIterator<Item = &'a SocketAddr>,
@@ -278,7 +277,7 @@ impl Sandbox {
         self.try_broadcast_to_addrs(msg, addresses).unwrap();
     }
 
-    // TODO: add self-test for broadcasting?
+    // TODO: Add self-test for broadcasting? (ECR-1627)
     pub fn try_broadcast_to_addrs<'a, T: Message, I>(
         &self,
         msg: &T,
@@ -470,10 +469,6 @@ impl Sandbox {
         let snapshot = self.blockchain_ref().snapshot();
         let schema = Schema::new(&snapshot);
         schema.actual_configuration()
-    }
-
-    pub fn propose_timeout(&self) -> Milliseconds {
-        self.node_state().propose_timeout()
     }
 
     pub fn majority_count(&self, num_validators: usize) -> usize {
@@ -694,8 +689,8 @@ pub fn sandbox_with_services_uninitialized(services: Vec<Box<Service>>) -> Sandb
         peers_timeout: 600_000,
         txs_block_limit: 1000,
         max_message_len: 1024 * 1024,
-        min_propose_timeout: 200,
-        max_propose_timeout: 200,
+        min_propose_timeout: PROPOSE_TIMEOUT,
+        max_propose_timeout: PROPOSE_TIMEOUT,
         propose_timeout_threshold: 0,
     };
     let genesis = GenesisConfig::new_with_consensus(
@@ -726,7 +721,7 @@ pub fn sandbox_with_services_uninitialized(services: Vec<Box<Service>>) -> Sandb
         mempool: Default::default(),
     };
 
-    // TODO use factory or other solution like set_handler or run
+    // TODO: Use factory or other solution like set_handler or run. (ECR-1627)
     let system_state = SandboxSystemStateProvider {
         listen_address: addresses[0],
         shared_time: SharedTime::new(Mutex::new(
@@ -772,7 +767,7 @@ pub fn sandbox_with_services_uninitialized(services: Vec<Box<Service>>) -> Sandb
     };
 
     // General assumption; necessary for correct work of consensus algorithm
-    assert!(sandbox.propose_timeout() < sandbox.round_timeout());
+    assert!(PROPOSE_TIMEOUT < sandbox.round_timeout());
     sandbox.process_events();
     sandbox
 }
@@ -880,7 +875,7 @@ mod tests {
 
     #[test]
     fn test_sandbox_assert_status() {
-        // TODO: remove this?
+        // TODO: Remove this? (ECR-1627)
         let s = timestamping_sandbox();
         s.assert_state(HEIGHT_ONE, ROUND_ONE);
         s.add_time(Duration::from_millis(999));