+++ exonum/src/storage/entry.rs
@@ -16,8 +16,7 @@
 
 use std::marker::PhantomData;
 
-use super::indexes_metadata::IndexType;
-use super::{BaseIndex, Fork, Snapshot, StorageValue};
+use super::{base_index::BaseIndex, indexes_metadata::IndexType, Fork, Snapshot, StorageValue};
 use crypto::Hash;
 
 /// An index that may only contain one element.