+++ exonum/src/storage/proof_list_index/proof.rs
@@ -12,15 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use serde::de::Error;
-use serde::ser::SerializeStruct;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{de::Error, ser::SerializeStruct, Deserialize, Deserializer, Serialize, Serializer};
 use serde_json::{from_value, Error as SerdeJsonError, Value};
 
 use self::ListProof::*;
-use super::super::StorageValue;
-use super::key::ProofListKey;
-use super::pair_hash;
+use super::{super::StorageValue, key::ProofListKey, pair_hash};
 use crypto::{hash, Hash};
 
 /// An enum that represents a proof of existence for a proof list elements.