+++ services/configuration/src/lib.rs
@@ -84,13 +84,13 @@ pub use errors::ErrorCode;
 pub use schema::{MaybeVote, ProposeData, Schema, VotingDecision};
 pub use transactions::{ConfigurationTransactions, Propose, Vote, VoteAgainst};
 
-use exonum::api::Api;
-use exonum::blockchain::{self, ApiContext, Transaction, TransactionSet};
-use exonum::crypto::Hash;
-use exonum::encoding::Error as EncodingError;
-use exonum::helpers::fabric::{self, Context};
-use exonum::messages::RawTransaction;
-use exonum::storage::Snapshot;
+use exonum::{api::Api,
+             blockchain::{self, ApiContext, Transaction, TransactionSet},
+             crypto::Hash,
+             encoding::Error as EncodingError,
+             helpers::fabric::{self, Context},
+             messages::RawTransaction,
+             storage::Snapshot};
 use iron::Handler;
 use router::Router;
 
@@ -103,6 +103,8 @@ mod transactions;
 
 /// Service identifier for the configuration service.
 pub const SERVICE_ID: u16 = 1;
+/// Configuration service name.
+pub const SERVICE_NAME: &str = "configuration";
 
 /// Configuration service.
 #[derive(Debug, Default)]
@@ -110,7 +112,7 @@ pub struct Service {}
 
 impl blockchain::Service for Service {
     fn service_name(&self) -> &'static str {
-        "configuration"
+        SERVICE_NAME
     }
 
     fn service_id(&self) -> u16 {
@@ -146,6 +148,10 @@ impl blockchain::Service for Service {
 pub struct ServiceFactory;
 
 impl fabric::ServiceFactory for ServiceFactory {
+    fn service_name(&self) -> &str {
+        SERVICE_NAME
+    }
+
     fn make_service(&mut self, _: &Context) -> Box<blockchain::Service> {
         Box::new(Service {})
     }