+++ services/time/src/lib.rs
@@ -44,18 +44,17 @@ pub mod time_provider;
 /// Node transactions.
 pub mod transactions;
 
-use exonum::api::Api;
-use exonum::blockchain::{ApiContext, Service, ServiceContext, Transaction, TransactionSet};
-use exonum::crypto::Hash;
-use exonum::encoding;
-use exonum::encoding::serialize::json::reexport::Value;
-use exonum::helpers::fabric::{Context, ServiceFactory};
-use exonum::messages::RawTransaction;
-use exonum::storage::{Fork, Snapshot};
-
+use exonum::{api::Api,
+             blockchain::{ApiContext, Service, ServiceContext, Transaction, TransactionSet},
+             crypto::Hash,
+             encoding::{self, serialize::json::reexport::Value},
+             helpers::fabric::{Context, ServiceFactory},
+             messages::RawTransaction,
+             storage::{Fork, Snapshot}};
 use iron::Handler;
 use router::Router;
 use schema::TimeSchema;
+
 use time_provider::{SystemTimeProvider, TimeProvider};
 use transactions::*;
 
@@ -156,6 +155,10 @@ impl Service for TimeService {
 pub struct TimeServiceFactory;
 
 impl ServiceFactory for TimeServiceFactory {
+    fn service_name(&self) -> &str {
+        SERVICE_NAME
+    }
+
     fn make_service(&mut self, _: &Context) -> Box<Service> {
         Box::new(TimeService::new())
     }