+++ testkit/tests/service_hooks/hooks.rs
@@ -14,12 +14,12 @@
 
 //! A special service which generates transactions on `after_commit` events.
 
-use exonum::blockchain::{ExecutionResult, Service, ServiceContext, Transaction, TransactionSet};
-use exonum::crypto::{Hash, Signature};
-use exonum::encoding;
-use exonum::helpers::Height;
-use exonum::messages::RawTransaction;
-use exonum::storage::{Fork, Snapshot};
+use exonum::{blockchain::{ExecutionResult, Service, ServiceContext, Transaction, TransactionSet},
+             crypto::{Hash, Signature},
+             encoding,
+             helpers::Height,
+             messages::RawTransaction,
+             storage::{Fork, Snapshot}};
 
 const SERVICE_ID: u16 = 512;